requirements.txt
functions-framework==3.*
google-cloud-bigquery>=3.11.0
google-cloud-logging
requests>=2.31.0
python-dateutil>=2.8.2
pydantic_settings>=2.12.0
pydantic>=2.12.5
facebook_business